+++ packages/components/CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
+#### Changed
+
+- `Text Field Wrapper` has improved RTL (right-to-left) support
+
 ### 3.8.0 - 2023-10-24
 
 ### 3.8.0-rc.0 - 2023-10-23