+++ packages/components/CHANGELOG.md
@@ -14,6 +14,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
+#### Fixed
+
+- `Checkbox Wrapper` and `Radio Button Wrapper` border-color/background-color does not reset on hover
+  ([#2852](https://github.com/porsche-design-system/porsche-design-system/pull/2852))
+
 ### 3.8.0 - 2023-10-24
 
 ### 3.8.0-rc.0 - 2023-10-23